Completely remove pyarrow.compat
